Moved Hexagons into separate component
To make it easier to use the hexagons in different pages, I extracted
the hexagon creation code and made a new Hexagon component that can be
imported into other parent components like Manual.

Added event dispatchers to Hexagon component that call functions in
parent component (ei: Manual) like buildCommandBlock for normal use and
testing.

Used bind:prop to communicate changes to props, like activeHexagon and
test_ui, in the parent component to Hexagon component.
