+++ src/routes/Manual.svelte
@@ -1,10 +1,10 @@
 <script>
 import { Router, Link, Route } from "svelte-routing";
-import { bind } from "svelte/internal";
+import Hexagons from "../Utils/Hexagons.svelte";
 
 // MARK: Message
 export let test_ui = false;
-
+$: pendingTimeout = false;
 $: message = 'starting ...';
 // $: message = message.toUpperCase();
 
@@ -105,7 +105,7 @@ $: lfPeriod = 500;
 $: lfDutyCycle = 50;
 $: hfPeriod = 250;
 $: hfDutyCycle = 50;
-$: single_pulse_duration = 500;
+$: single_pulse_duration = 500; //low priority
 
 
 $: single_pulse_block = [0,0,0,0];
@@ -129,9 +129,7 @@ function setTimingBlock(config) {
 
 // MARK: Device Input
 
-$: activeHexagon = -1;
-$: mouseDown = false;
-$: pendingTimeout = false;
+let activeHexagon = -1;
 
 let block0_31 = [0,0,0,0];
 let block32_63 = [0,0,0,0];
@@ -154,106 +152,6 @@ $: act_command = `{ "ActuatorsCommand": {
       "lf_block":{"b0":${lf_block[0]}, "b1":${lf_block[1]}, "b2":${lf_block[2]}, "b3":${lf_block[3]}}
     } } }`;
 
-
-
-const hexagons = [
-  // Row 0
-  { id: 0, row: 0, col: 1 },
-  { id: 1, row: 0, col: 3 },
-  // filler
-  { id: 2, row: 0, col: 7 },
-  { id: 3, row: 0, col: 9 },
-
-  // Row 1
-  { id: 4, row: 1, col: 0 },
-  { id: 5, row: 1, col: 2 },
-  { id: 6, row: 1, col: 4 },
-  { id: 7, row: 1, col: 6 },
-  { id: 8, row: 1, col: 8 },
-  { id: 9, row: 1, col: 10 },
-
-  // Row 2
-  { id: 10, row: 2, col: 1 },
-  { id: 11, row: 2, col: 3 },
-  { id: 12, row: 2, col: 5 },
-  { id: 13, row: 2, col: 7 },
-  { id: 14, row: 2, col: 9 },
-
-  // Row 3
-  { id: 15, row: 3, col: 0 },
-  { id: 16, row: 3, col: 2 },
-  { id: 17, row: 3, col: 4 },
-  { id: 18, row: 3, col: 6 },
-  { id: 19, row: 3, col: 8 },
-  { id: 20, row: 3, col: 10 },
-
-  // Row 4
-  { id: 21, row: 4, col: 1 },
-  { id: 22, row: 4, col: 3 },
-  { id: 23, row: 4, col: 5 },
-  { id: 24, row: 4, col: 7 },
-  { id: 25, row: 4, col: 9 },
-
-  // Row 5
-  { id: 26, row: 5, col: 0 },
-  { id: 27, row: 5, col: 2 },
-  { id: 28, row: 5, col: 4 },
-  { id: 29, row: 5, col: 6 },
-  { id: 30, row: 5, col: 8 },
-  { id: 31, row: 5, col: 10 },
-
-  // Row 6
-  { id: 32, row: 6, col: 1 },
-  { id: 33, row: 6, col: 3 },
-  // filler
-  { id: 34, row: 6, col: 7 },
-  { id: 35, row: 6, col: 9 },];
-
-const hexagonSideLength = 70;
-const globalPadding = 10;
-const baseSpacing = 15;
-const horizontalSpacing = baseSpacing;
-const verticalSpacing = 1.5 * baseSpacing;
-const width = Math.sqrt(3) * hexagonSideLength;
-const height = 2 * hexagonSideLength;
-
-// Compute the pixel location for each hexagon
-const hexagonsWithoutPixels = hexagons.map(({id, row, col}) => {
-    return {
-    id,
-    row,
-    col,
-    x: (globalPadding / 2) + (col/2 + .5) * width + Math.max((col / 2) * horizontalSpacing, 0),
-    y: (globalPadding / 2) + ((row/2) * 1.5 + .5) * height + Math.max((row / 2) * verticalSpacing, 0),
-    }
-});
-
-const viewBox = hexagonsWithoutPixels.reduce((previousValue, currentValue) => {
-    return {
-        x: Math.max(previousValue.x, currentValue.x + (width / 2) + (globalPadding / 2)),
-        y: Math.max(previousValue.y, currentValue.y + (height / 2) + (globalPadding / 2)),
-    };
-});
-
-$: drawableHexagons = hexagons.map(({id, row, col}) => {
-    const sqrt3 = Math.sqrt(3);
-    const a = hexagonSideLength / 2;
-    const height = (4 * a);
-    const width = (2 * sqrt3 * a);
-    const points = [
-      `${sqrt3 * a}, ${0}`,
-      `${width}, ${a}`,
-      `${width}, ${3 * a}`,
-      `${sqrt3 * a}, ${height}`,
-      `${0}, ${3 * a}`,
-      `${0}, ${a}`,
-    ];
-    const x = (globalPadding / 2) + (col/2 + .5) * width + Math.max((col / 2) * horizontalSpacing, 0);
-    const y = (globalPadding / 2) + ((row/2) * 1.5 + .5) * height + Math.max((row / 2) * verticalSpacing, 0);
-    const color = activeHexagon === id ? "mediumseagreen" : "black";
-
-    return {id, x, y, width, height, points: points.join(' '), color};
-});
 
 function getBytesForActuator(actuator) {
     /*Passed active actuator, Returns Array of 16 strings with 8 'bits'*/
@@ -289,59 +187,23 @@ function buildCommandBlocks(active) {
 
 }
 
-
-function handleTouchStart(e) { 
-    if(e.stopPropagation) e.stopPropagation();
-    if(e.preventDefault) e.preventDefault();
-    mouseDown = true;
+//Handles mouseEvents fomr hexgon component
+function handleActuatorClick(e) {
+    buildCommandBlocks(e.detail.value);
+    (async () => {
+        return await hitEndpoint(endpoint, nopRoute, act_command);
+    })().then(result => {
+        if(pendingTimeout) {
+            clearTimeout(pendingTimeout);
+        }
+        pendingTimeout = setTimeout(() => { activeHexagon = -1 }, 500);
+        return result;
+    }).catch(error => {
+        message = error;
+        activeHexagon = -1;
+        throw error;
+    });
 }
-function handleTouchMove(e) {
-    if(e.stopPropagation) e.stopPropagation();
-    if(e.preventDefault) e.preventDefault();
-
-    if(!mouseDown) {
-        return;
-    }
-
-    const radius = Math.sqrt(3) * hexagonSideLength / 2;
-    const radiusSquared = radius * radius;
-
-    for(var i = 0; i < drawableHexagons.length; i++) {
-        const x = drawableHexagons[i].x;
-        const y = drawableHexagons[i].y;
-        const xDist = (x - e.offsetX);
-        const yDist = (y - e.offsetY);
-        const euclidianDistSquared = xDist * xDist + yDist * yDist;
-        if(euclidianDistSquared < radiusSquared) {
-            activeHexagon = drawableHexagons[i].id;
-            buildCommandBlocks(activeHexagon);
-            (async () => {
-               return await hitEndpoint(endpoint, nopRoute, act_command);
-            })().then(result => {
-                if(pendingTimeout) {
-                    clearTimeout(pendingTimeout);
-                }
-                pendingTimeout = setTimeout(() => { activeHexagon = -1 }, 500);
-                return result;
-            }).catch(error => {
-                message = error;
-                activeHexagon = -1;
-                throw error;
-            });
-      }
-    }
-}
-function handleTouchEnd(e) {
-    if(e.stopPropagation) e.stopPropagation();
-    if(e.preventDefault) e.preventDefault();
-
-    mouseDown = false;
-    activeHexagon = -1;
-}
-
-function handleMouseDown(event) { handleTouchStart(event); }
-function handleMouseMove(event) { handleTouchMove(event); }
-function handleMouseUp(event) { handleTouchEnd(event); }
 
 </script>
 
@@ -442,42 +304,11 @@ function handleMouseUp(event) { handleTouchEnd(event); }
                     
             </Route>
         </div>
-        <div class='col-50'>
-            <h2>Device Input</h2>
+        <Hexagons on:click={handleActuatorClick} on:message="{e => buildCommandBlocks(e.detail.value)}" bind:activeHexagon bind:test_ui/>
 
-            <svg version="1.0" xmlns="http://www.w3.org/2000/svg"
-                width={`${viewBox.x}px`}
-                height={`${viewBox.y}px`}
-                viewBox={`0 0 ${viewBox.x} ${viewBox.y}`}
-                preserveAspectRatio="xMidYMid meet"
-                on:mousedown={handleMouseDown}
-                on:mousemove={handleMouseMove}
-                on:mouseup={handleMouseUp}
-                >
-
-                {#each drawableHexagons as hexagon}
-                    <g key={`g-${hexagon.id}`}>
-                        <polygon
-                          id={`hex-${hexagon.id}`}
-                          key={`hex-${hexagon.x}-${hexagon.y}`}
-                          class="activatable"
-                          fill="none"
-                          stroke={hexagon.color}
-                          stroke-width="5px"
-                          strokeLinejoin="miter"
-                          transform={`rotate(0 ${hexagon.x} ${hexagon.y}) translate(${hexagon.x - (hexagon.width / 2)} ${hexagon.y - (hexagon.height / 2)})`}
-                          points={hexagon.points} />
-                          <text id={`text-${hexagon.id}`} x={hexagon.x - 6} y={hexagon.y}>{hexagon.id}</text>
-                      </g>
-                {/each}
-            </svg>
-            {#if test_ui}
+        {#if test_ui}
                 <p>Test: {block0_31}, {block32_63}, {block64_95}, {block96_127}</p>
-                {#each hexagons as actuator}
-                    <button on:click={() => buildCommandBlocks(actuator.id)}> Button {actuator.id}</button>
-                {/each}
-            {/if}
-        </div>
+        {/if}
   </div>
 
 </Router>
@@ -500,7 +331,7 @@ function handleMouseUp(event) { handleTouchEnd(event); }
         padding: 1em;
     }
 
-    .col-50 {
+    /* .col-50 {
         width: 40%;
         height: 100%;
 
@@ -509,7 +340,7 @@ function handleMouseUp(event) { handleTouchEnd(event); }
 
         margin: 0px auto;
         padding: 1em;
-    }
+    } */
 
     p.success {
         color: green;
@@ -525,9 +356,9 @@ function handleMouseUp(event) { handleTouchEnd(event); }
         margin: 1em;
     }
 
-    svg {
+    /* svg {
         cursor: draggable;
-    }
+    } */
 
 </style>
